Move sort code to new folder
