+++ NuKeeper/Engine/Packages/PackageUpdateSelection.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using NuKeeper.Configuration;
+using NuKeeper.Engine.Sort;
 using NuKeeper.Inspection.Logging;
 using NuKeeper.Inspection.RepositoryInspection;
