+++ NuKeeper/Engine/Sort/TopologicalSort.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using NuKeeper.Inspection.Logging;
 using NuKeeper.Inspection.RepositoryInspection;
 
-namespace NuKeeper.Engine.Packages
+namespace NuKeeper.Engine.Sort
 {
     /// <summary>
     /// https://en.wikipedia.org/wiki/Topological_sorting#Depth-first_search
